SuperWorm 1.6
Added Game Difficulty as well as styling for each button press.
Reduced the amount of clutter outside the game and either put in in the
start menu or the game itself.
Removed useless CSS.
Added Instructions inside the start menu.
Added the ability to return to start screen when paused to change volume
or themes.
Added more hotkeys to navigate quicker.
Added some styling to make the game look a little cleaner.
Modified the score calculation and included difficulty bonuses.